Parallel interpolation of climatological monthly albedo (albedo12m)
This commit adds logic to interpolate the climatological monthly albedo
(albedo12m) safely in parallel on a non-cvt partition. Although all fields now
have logic to safely interpolate static fields in parallel on a non-cvt
decomposition, the check in mpas_init_cases is still present so processing in
parallel is currently not possible.